+++ CHANGELOG.md
@@ -3,8 +3,10 @@
 ## [Unreleased]
 
 - Added throw on missing/invalid state
+- Added extra params option to login method
 - Changed lock implementation to use broadcast channel in browsers than support it
 - Removed the 0-1s refresh delay that is no longer needed
+- Removed downlevel ts
 
 ## [0.2.0] - 2021-07-09